fix: ingress watcher does not work
In order to reduce number of reloads, the ingress watch loop
implemented a draining mechanism on the watch channel that tried to
consume all pending events on the channel after a config reload. The
actual implementation of this draining mechanism however, was not
working correctly.  In order to fix the problem, a debounce mechanism
was implemented, that debounces events on the watch channel.